% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_normalize.R
\name{npi_normalize}
\alias{npi_normalize}
\title{Normalize user input rows for lake search & matching}
\usage{
npi_normalize(inputs, full_name = "full_name", state = "state", city = "city")
}
\arguments{
\item{inputs}{data.frame/tibble with at least full_name + state (and optional city)}

\item{full_name}{Column name for full name}

\item{state}{Column name for state (2-letter preferred)}

\item{city}{Column name for city (optional)}
}
\description{
Normalize user input rows for lake search & matching
}
