% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{tnp_duckdb}
\alias{tnp_duckdb}
\title{Open a DuckDB connection configured for HTTP(S) parquet}
\usage{
tnp_duckdb(dbdir = ":memory:")
}
\arguments{
\item{dbdir}{Path to DuckDB database directory (default ":memory:").}
}
\value{
A DBI connection to DuckDB with httpfs loaded.
}
\description{
Open a DuckDB connection configured for HTTP(S) parquet
}
