% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_search.R
\name{tnp_lake_read}
\alias{tnp_lake_read}
\title{Read a slice from the lake}
\usage{
tnp_lake_read(
  con,
  man,
  state,
  initials = NULL,
  where_sql = NULL,
  columns = c("npi", "first_name", "last_name", "state")
)
}
\arguments{
\item{con}{duckdb connection}

\item{man}{manifest list}

\item{state}{2-letter code (e.g., "NY")}

\item{initials}{optional initials vector to prune shards}

\item{where_sql}{optional SQL WHERE clause to apply server-side}

\item{columns}{optional columns to select (default some key fields)}
}
\description{
Read a slice from the lake
}
