% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{tnp_norm_name}
\alias{tnp_norm_name}
\title{Normalize simple person names}
\usage{
tnp_norm_name(x)
}
\arguments{
\item{x}{Character vector of names to normalize.}
}
\value{
Uppercased, ASCII-like, squished name strings.
}
\description{
Normalize simple person names
}
