% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{tnp_state_initials}
\alias{tnp_state_initials}
\title{List available initials for a state}
\usage{
tnp_state_initials(man, state)
}
\arguments{
\item{man}{Manifest list returned by \code{\link[=tnp_manifest]{tnp_manifest()}}.}

\item{state}{Two-letter state code (e.g., "NY").}
}
\value{
A character vector of initials present in the lake for that state.
}
\description{
List available initials for a state
}
