% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tnp_taxonomy_dict}
\alias{tnp_taxonomy_dict}
\title{Download and cache the NUCC taxonomy dictionary}
\usage{
tnp_taxonomy_dict(url = tnp_taxonomy_url())
}
\arguments{
\item{url}{Optional override URL}
}
\description{
Download and cache the NUCC taxonomy dictionary
}
