% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tnp_taxonomy_translate}
\alias{tnp_taxonomy_translate}
\title{Translate taxonomy code columns into readable labels}
\usage{
tnp_taxonomy_translate(x, cols = NULL)
}
\arguments{
\item{x}{A data.frame/tibble with columns like tax_code_1, tax_code_2, ...}

\item{cols}{Which columns to translate (default: all columns starting with "tax_code_")}
}
\description{
Translate taxonomy code columns into readable labels
}
